Fix test: wait for promise resolution before unmount
The 'disables buttons while creating' test was causing an unhandled rejection
because it used setTimeout and didn't wait for the promise to complete before
unmounting the component. This caused setIsCreating(false) to be called after
the component was unmounted.

Fixed by using a manually-resolvable promise instead of setTimeout, and waiting
for the promise to complete before the test ends.